backend/foodgram_django/requirements.txt
Django==2.2.16
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
